fixed vote, added isAdmin

+++ src/main/resources/app/components/petition/main/petition.main.component.ts
@@ -3,6 +3,7 @@ import { HttpService } from "./../../../services/httpServices/http.service";
 import { Petition } from "./../../../model/Petition";
 import {ActivatedRoute, Params, Router} from "@angular/router";
 import {ShortPetition} from "../../../model/ShortPetition";
+import {AuthenticationService} from "../../../services/httpServices/authenticationServices/authentication.service";
 
 @Component({
     templateUrl: 'app/components/petition/main/petition.main.component.html',
@@ -11,7 +12,7 @@ import {ShortPetition} from "../../../model/ShortPetition";
 export class PetitionMainComponent implements OnInit {
     private petition: Petition = new Petition();
 
-    constructor(private http: HttpService, private activatedRoute: ActivatedRoute, private router: Router) { }
+    constructor(private http: HttpService, private activatedRoute: ActivatedRoute, private router: Router, private authService: AuthenticationService) { }
 
 	ngOnInit() {
 
@@ -27,13 +28,16 @@ export class PetitionMainComponent implements OnInit {
 	}
 
 	delete(): void {
-		this.activatedRoute.params.subscribe((params: Params) => {
-			let petitionId = params['id'];
-			this.http
-                .deleteData("/petition/" + petitionId)
-                .subscribe(data => {
-                	this.router.navigate(["/petitions"]);
-				});
-		});
+		var isDelete = confirm("Delete?");
+		if(isDelete) {
+			this.activatedRoute.params.subscribe((params: Params) => {
+				let petitionId = params['id'];
+				this.http
+					.deleteData("/petition/" + petitionId)
+					.subscribe(data => {
+						this.router.navigate(["/petitions"]);
+					});
+			});
+		}
 	}
 }
